New logs, and bunch of other stuff. This will be first version to release
