+++ .github/workflows/maven.yml
@@ -11,7 +11,10 @@ name: Java CI with Maven
 on:
   push:
     branches:
-      - dev
+      - main
+  pull_request:
+    branches:
+      - main
       
 jobs:
   buildRelese:
@@ -37,8 +40,8 @@ jobs:
       uses: "marvinpinto/action-automatic-releases@latest"
       with:
         repo_token: "${{ secrets.GITHUB_TOKEN }}"
-        automatic_release_tag: "pre-${{ github.sha }}"
-        prerelease: true
-        title: "Development Build pre-${{ github.sha }}"
+        automatic_release_tag: "b-${{ github.sha }}"
+        prerelease: false
+        title: "Development Build ${{ github.sha }}"
         files: |
           ./target/${{ vars.JAR_NAME }}.jar
