+++ commands.md
@@ -1,0 +1,37 @@
+
+## Categories
+### 1. Gamemodes
+```
+/gmc /gms /gmsp /gmc
+/gamemode|gm <gamemode> [player]
+```
+### 2. Teleportation
+```
+/setspawn 
+/spawn
+```
+### 3. Inventory
+```
+/invsee <player>
+/workbench /craft /wb
+/anvil
+/enchanttable /et
+/enderchest /ec
+```
+### 4. Chat
+```
+/cmdspy
+/clearchat
+/lockchat
+```
+### 5. Utils
+```
+/feed /heal /fly /speed
+/whois /ping
+/msg <player> <message>
+/r <message>
+```
+### 6. Weather
+```
+/day /night /sun /rain /thunder
+```
