+++ readme.md
@@ -1,30 +1,66 @@
 # XaCore
 
 ## Description
-Something like essentialsX plugin but a lot smaller.
+Plugin that contains basic commands and features for a Minecraft server. <br>
+Supports Vault and PlaceholderAPI. <br>
 
-## Commands
-- /spawn - /setspawn
-- /home - /sethome
-- /anvil, /craft, /enderchest, /enchanttable
-- /feed, /heal
-- /fly, /speed
-- /clearchat, /lockchat
-- /day, /night, /sun, /rain, /thunder
-- /whois, /ping
-- /gm, /gmc, /gms, /gmsp, /gma
-- /invsee
+<details>
+    <summary><b>Commands categories</b></summary>
 
-## Features
+#### 1. Gamemodes
+```
+/gmc /gms /gmsp /gmc
+/gamemode|gm <gamemode> [player]
+```
+#### 2. Teleportation
+```
+/setspawn 
+/spawn
+```
+#### 3. Inventory
+```
+/invsee <player>
+/workbench /craft /wb
+/anvil
+/enchanttable /et
+/enderchest /ec
+```
+#### 4. Chat
+```
+/cmdspy
+/clearchat
+/lockchat
+```
+#### 5. Utils
+```
+/feed /heal /fly /speed
+/whois /ping
+/msg <player> <message>
+/r <message>
+```
+#### 6. Weather
+```
+/day /night /sun /rain /thunder
+```
+</details>
+<br>
+<details>
+    <summary><b>Features</b></summary>
+
 - Supporting PAPI placeholders in all messages (in lang.yml)
 - Customizable join/quit messages and chat format
 - Command blacklist
-- Spawn on join
+- Spawn on join, gamemode on join
+- Customizable world flags like: disable fall damage, block break etc..
 
-## TODO
+</details>
+<br>
+<details>
+    <summary><b>TODO</b></summary>
+
 - MOTD
 - Customizable tablist
 - Customizable scoreboard
-- NoFall, anti-hunger, other flags per worlds
 - Tpa, tpahere, tpaccept, tpdeny
 - Warps, homes
+</details>
