+++ src/main/java/me/xap3y/xacore/Main.kt
@@ -1,16 +1,15 @@
 package me.xap3y.xacore
 
-import me.xap3y.xacore.api.config.ConfigManager
+import me.xap3y.xacore.api.config.StorageManager
+import me.xap3y.xacore.api.enums.CommandCategory
 import me.xap3y.xacore.api.text.Texter
 import me.xap3y.xacore.commands.*
-import me.xap3y.xacore.listeners.PlayerChatListener
-import me.xap3y.xacore.listeners.PlayerCommandPreprocessListener
-import me.xap3y.xacore.listeners.PlayerJoinListener
-import me.xap3y.xacore.listeners.PlayerQuitListener
+import me.xap3y.xacore.listeners.*
 import me.xap3y.xacore.utils.Helper
 import me.xap3y.xacore.utils.HookManager
 import net.milkbowl.vault.chat.Chat
 import net.milkbowl.vault.permission.Permission
+import org.apache.logging.log4j.CloseableThreadContext.Instance
 import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
 import org.bukkit.event.Listener
@@ -24,20 +23,21 @@ import java.io.File
 
 class Main : JavaPlugin() {
 
-    lateinit var configManager: ConfigManager
+    lateinit var storageManager: StorageManager
     lateinit var configFile: File
     lateinit var messageFile: File
-
-    var chatLocked = false
-    val textApi: Texter by lazy { Texter(this) }
-    val helper: Helper by lazy { Helper(this) }
-
-    val whisperPlayers: HashMap<CommandSender, CommandSender> = hashMapOf()
-
+    lateinit var logFile: File
     lateinit var vaultPair: Pair<Chat, Permission>
 
+    var chatLocked = false
     var useVault = false
     var usePapi = false
+
+    val helper: Helper by lazy { Helper(this) }
+    val textApi: Texter by lazy { Texter(this) }
+    val whisperPlayers: HashMap<CommandSender, CommandSender> = hashMapOf()
+    val cmdSpyToggles: MutableSet<Player> = mutableSetOf()
+
 
     override fun onEnable() {
 
@@ -46,21 +46,34 @@ class Main : JavaPlugin() {
 
         configFile = File(dataFolder, "config.yml")
         messageFile = File(dataFolder, "lang.yml")
+        logFile = File(dataFolder, "logs.txt")
 
-        configManager = ConfigManager(this)
-        configManager.loadConfig()
-        configManager.loadLang()
+        storageManager = StorageManager(this)
+        storageManager.loadConfig()
+        storageManager.loadLang()
 
         useVault = config.getBoolean("hookVault")
 
         val commandManager = createCommandManager()
         val annotationParser = createAnnotationParser(commandManager)
 
-        annotationParser.parse(RootCommand(this))
-        annotationParser.parse(Gamemodes(this))
-        annotationParser.parse(WeatherCommands(this))
-        annotationParser.parse(AdminCommands(this))
-        annotationParser.parse(UtilityCommands(this))
+
+        val commandCategories: HashMap<CommandCategory, Any> =
+            hashMapOf(
+                CommandCategory.GAMEMODE to GamemodeCommands(this),
+                CommandCategory.WEATHER to WeatherCommands(this),
+                CommandCategory.CHAT to ChatCommands(this),
+                CommandCategory.INVENTORY to InventoryCommands(this),
+                CommandCategory.UTILS to UtilityCommands(this),
+                CommandCategory.TELEPORT to TeleportationCommands(this)
+            )
+
+        val list = config.getList("commandCategoriesDisabled") ?: listOf()
+
+        commandCategories.forEach { (k, v) ->
+            if (list.contains(k.name.lowercase())) return@forEach
+            annotationParser.parse(v)
+        }
 
         registerListeners()
 
@@ -83,11 +96,20 @@ class Main : JavaPlugin() {
                 if (usePapi) usePapi = false
         }
 
+        val time = System.currentTimeMillis()
+        val date = java.util.Date(time)
+        val humanDate = date.toString()
+
+        storageManager.logInfo("Plugin enabled at $humanDate")
 
     }
 
     override fun onDisable() {
-        // Plugin shutdown logic
+        val time = System.currentTimeMillis()
+        val date = java.util.Date(time)
+        val humanDate = date.toString()
+
+        storageManager.logInfo("Plugin disabled at $humanDate")
     }
 
     private fun registerListeners() {
@@ -95,7 +117,11 @@ class Main : JavaPlugin() {
             PlayerJoinListener(this),
             PlayerQuitListener(this),
             PlayerChatListener(this),
-            PlayerCommandPreprocessListener(this)
+            PlayerCommandPreprocessListener(this),
+            EntityDamageListener(this),
+            BlockListener(this),
+            ServerCommandListener(this.storageManager),
+            FoodLevelChangeListener(this)
         )
 
         val pm: PluginManager = server.pluginManager
@@ -134,4 +160,5 @@ class Main : JavaPlugin() {
     fun isConfigFileInit() = ::configFile.isInitialized
     fun isLangFileInit() = ::messageFile.isInitialized
     fun isVaultPairInit() = ::vaultPair.isInitialized
+    fun isLogFileInit() = ::logFile.isInitialized
 }
