+++ src/main/java/me/xap3y/xacore/api/enums/CommandCategory.kt
@@ -1,0 +1,10 @@
+package me.xap3y.xacore.api.enums
+
+enum class CommandCategory {
+    CHAT,
+    WEATHER,
+    GAMEMODE,
+    TELEPORT,
+    INVENTORY,
+    UTILS
+}
