+++ src/main/java/me/xap3y/xacore/listeners/FoodLevelChangeListener.kt
@@ -1,0 +1,21 @@
+package me.xap3y.xacore.listeners
+
+import me.xap3y.xacore.Main
+import org.bukkit.entity.Player
+import org.bukkit.event.EventHandler
+import org.bukkit.event.Listener
+import org.bukkit.event.entity.FoodLevelChangeEvent
+
+class FoodLevelChangeListener(private val plugin: Main): Listener {
+
+    @EventHandler
+    fun onHungerChange(e: FoodLevelChangeEvent) {
+        val player = e.entity as Player
+        val interrupt = plugin.helper.interruptEventOnFlag(player.world.name, "noHunger")
+
+        if (interrupt){
+            if (e.foodLevel < 20) e.foodLevel = 20
+            e.isCancelled = true
+        }
+    }
+}
