+++ src/main/java/me/xap3y/xacore/listeners/PlayerJoinListener.kt
@@ -13,9 +13,9 @@ class PlayerJoinListener(private val plugin: Main): Listener {
     fun onPlayerJoinEvent(e: PlayerJoinEvent) {
 
         // Custom join message
-        val joinMessageEnabled = plugin.config.getBoolean("joinMessage")
+        val joinMessageEnabled = plugin.config.getBoolean("joinMessage", false)
         if (joinMessageEnabled) {
-            val message = plugin.configManager.getMessage("messages.joinMessage", "&6<player> &fjoined the game", e.player)
+            val message = plugin.storageManager.getMessage("messages.joinMessage", "&6<player> &fjoined the game", e.player)
 
             e.joinMessage = plugin.textApi.coloredMessage(
                 plugin.textApi.replace(
@@ -27,7 +27,7 @@ class PlayerJoinListener(private val plugin: Main): Listener {
         }
 
         // Spawn on player join
-        val spawnOnJoinEnabled = plugin.config.getBoolean("spawnOnJoin")
+        val spawnOnJoinEnabled = plugin.config.getBoolean("spawnOnJoin", false)
 
         if (spawnOnJoinEnabled) {
 
@@ -45,7 +45,7 @@ class PlayerJoinListener(private val plugin: Main): Listener {
         // TODO -- Scoreboard and maybe tab-list?
 
         // Gamemode on join
-        val gamemodeOnJoinToggle = plugin.config.getBoolean("gamemodeOnJoinToggle")
+        val gamemodeOnJoinToggle = plugin.config.getBoolean("gamemodeOnJoinToggle", false)
         if (gamemodeOnJoinToggle) {
 
             val gamemodeOnJoin = plugin.config.getString("gamemodeOnJoin") ?: "SURVIVAL"
